refactoring code to point to the shared and business sub dirs
