+++ shared/src/business/entities/Case.js
