+++ shared/src/business/entities/Case.test.js
