+++ shared/src/business/entities/CaseInitiator.js
