+++ shared/src/business/entities/Document.js
