+++ shared/src/business/entities/User.js
