+++ shared/src/business/useCases/getCaseProxy.js
