+++ shared/src/business/useCases/getCasesByStatus.js
@@ -1,8 +1,8 @@
 const {
   isAuthorized,
   GET_CASES_BY_STATUS,
-} = require('../authorization/authorizationClientService');
-const { UnauthorizedError } = require('../errors/errors');
+} = require('../../authorization/authorizationClientService');
+const { UnauthorizedError } = require('../../errors/errors');
 
 /**
  * getCasesByStatus
