+++ shared/src/business/useCases/getCasesByStatusProxy.js
