+++ shared/src/business/useCases/getCasesByUser.js
