+++ shared/src/business/useCases/getCasesByUserProxy.js
