+++ shared/src/business/useCases/getCasesProxy.js
