+++ shared/src/business/useCases/getUser.js
