+++ web-client/src/applicationContexts/dev.js
@@ -1,12 +1,12 @@
 import persistenceGateway from '../../../shared/src/persistence/awsPersistenceGateway';
 
-import createCase from '../../../shared/src/useCases/createCaseProxy';
-import getCase from '../../../shared/src/useCases/getCaseProxy';
-import getCasesByStatus from '../../../shared/src/useCases/getCasesByStatusProxy';
-import getCasesByUser from '../../../shared/src/useCases/getCasesByUserProxy';
-import getUser from '../../../shared/src/useCases/getUser';
-import updateCase from '../../../shared/src/useCases/updateCaseProxy';
-import uploadCasePdfs from '../../../shared/src/useCases/uploadCasePdfs';
+import createCase from '../../../shared/src/business/useCases/createCaseProxy';
+import getCase from '../../../shared/src/business/useCases/getCaseProxy';
+import getCasesByStatus from '../../../shared/src/business/useCases/getCasesByStatusProxy';
+import getCasesByUser from '../../../shared/src/business/useCases/getCasesByUserProxy';
+import getUser from '../../../shared/src/business/useCases/getUser';
+import updateCase from '../../../shared/src/business/useCases/updateCaseProxy';
+import uploadCasePdfs from '../../../shared/src/business/useCases/uploadCasePdfs';
 
 /**
  * Context for the dev environment
